+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
-- NA
+- Added [Tee-Clip] (Alias: tclip) function
 
 ## [7.3.0] - 2024-10-24
 
@@ -770,7 +770,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Add-ID]: src/Add-ID_function.ps1
 
-[Add-ID]: src/Join-Line2_function.ps1
+[Join-Line2]: src/Join-Line2_function.ps1
+
+[Tee-Clip]: src/Tee-Clip_function.ps1
 
 [unreleased]: https://github.com/btklab/posh-mocks/compare/7.3.0..HEAD
 [7.3.0]: https://github.com/btklab/posh-mocks/releases/tag/7.3.0